# api-service/docs/swagger.yaml
definitions:
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  api.LoginResponse:
    properties:
      token:
        type: string
    type: object
  api.RegisterRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  api.ResultResponse:
    properties:
      result:
        type: string
    type: object
  api.StatusResponse:
    properties:
      status:
        type: string
    type: object
  api.TaskResponse:
    properties:
      task_id:
        type: string
    type: object
  model.Task:
    properties:
      code:
        type: string
    type: object
info:
  contact: {}
  title: Orders API
paths:
  /login:
    post:
      consumes:
      - application/json
      description: perform authorization and return session`s token if successful
      parameters:
      - description: username + password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unathorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "405":
          description: Method not allowed
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: login user
      tags:
      - auth
      - user
  /register:
    post:
      consumes:
      - application/json
      description: Register new user and return token
      parameters:
      - description: username + password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "405":
          description: Method not allowed
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: failed to create session
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Register new user and return token
      tags:
      - auth
      - user
  /result/{task_id}:
    get:
      consumes:
      - application/json
      description: Return result of task by id
      parameters:
      - description: session`s token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Task id
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: return result
          schema:
            $ref: '#/definitions/api.ResultResponse'
        "202":
          description: task is running
          schema:
            $ref: '#/definitions/api.ResultResponse'
        "401":
          description: Unathorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Result not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "405":
          description: Method not allowed
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Return result of task by id
      tags:
      - task
  /status/{id}:
    get:
      consumes:
      - application/json
      description: Return status of task by id
      parameters:
      - description: session`s token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "401":
          description: Unathorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Status not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "405":
          description: Method not allowed
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Return status of task by id
      tags:
      - task
  /task:
    post:
      consumes:
      - application/json
      description: Create new task and return task`s id
      parameters:
      - description: session's token
        in: header
        name: Authorization
        required: true
        type: string
      - description: code to run
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TaskResponse'
        "401":
          description: Unathorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "405":
          description: Method not allowed
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create new task
      tags:
      - task
swagger: "2.0"